fixing mess with same names for different partticaonts #1
Initiate exploration of strategies to correct file naming inconsistencies due to RA errors. This commit focuses on developing a methodical approach to detect, report, and correct instances where participant identifiers have been reused or incorrectly labeled across our dataset. This is crucial for ensuring data integrity and reliability, especially for identifiers 118, 119, 222, and 223, which have been identified as problematic. The aim is to automate as much of this correction process as possible, reducing the need for manual intervention and decreasing the likelihood of similar issues arising in the future. Details on the proposed solutions and their implementation will be shared in subsequent updates. #1